refactor(frontend): extract onboarding module
Load onboarding only for first-time users and inject UI dependencies to avoid a module cycle.
